Simplify bootstrapping
